Allow to make the project, make examples and run them.
The makefile looks for .cpp files in the ../src directory to produce .o files and link them into the tracer pass.
Also, all .c files in the ../bench directory can be transformed into the appropriate .ll files.
Use tracer.sh script to run the tracer pass on the .ll file.
Note: this is only framework yet, so it produces no real transformations.