Complete Alloca insrumentation for both Array and Integer types
Alloca handling strategy is made in uniform way for arrays and integers
- no pre-defined type specifier constants are needed.
First argument in alloca_handler state the same - this is a pointer to the allocated memory.
Second argument is a pointer to the string representing an appropriate type
(as it is printed by LLVM 'operator<<' on the 'Type*' object to the raw_string_stream).
Memoization is implemented - no duplicate type specifiers could occure in the transformed program.