Extract lazy, memoized string buffers initialization from alloca transformation pass
For every alloca instruction transformed program needs to pass type of the allocated object to the appropriate handler as an argument.
But not only alloca needs to store and pass such information as strings - call and ret instructions also need such mechanism
to pass function names to the handlers. Also, alloca needs to pass variable name to the appropriate handler.
For such reason string buffers initialization mechanism was extracted from the alloca pass and placed to a separate unit.
This initializatio is LAZY because it initializes string buffers only when it is needed.
It also MEMOIZE because it mantains only ONE buffer for ALL equvalent strings.