Fix engine initialization bug
There was a bug in the Engine initialization. The Engine used to push new stack frame on each alloca instruction handling call!
To fix this nonsense behavior Engine constructor was implemented and appropriate stack initialization moved to the Engine's constructor.
Note: maybe it would be useful to move this initialization furhter, to the FrameStack constructor.